add unit tests for the CompareSAX test utility
** About running the new tests with disabled Python assertions **

Since the CompareSAX uses pytest assertions, it can not be tested when Python
assertions have been disabled in that module, but pytest assertion rewriting
saves us from this if it is enabled. This effectively means that the new tests
can not be run with disabled Python assertions without pytest's assertion
rewriting being enabled.

Since pytest assertion rewriting has not been implemented for Python versions
prior to 2.6, the new tests can not be run with disabled Python assertions using
such Python versions at all.